+++ docs/src/history.md
@@ -4,6 +4,7 @@
 ## Current development version
 
 * CHANGED: `getindex()` changed to `get()`, since we need a default value to be provided.
+* ADDED: `Selectable` abstract type to restrict the usage of constant-time `get()`.
 
 
 ## v0.1.0